* restructure cypress code * add dummy integration, non-cucumber test
