+++ cypress/e2e/Authentication/LogConsumerOut/main.js
@@ -6,7 +6,7 @@ import {
   getTokenCookie,
   checkIfLinkIsActive
 } from '../../../../common/cypress/integration/Authentication/common/Helpers'
-import '../../../../common/cypress/integration/Authentication/common/PersonaType'
+import '../../../../common/cypress/integration/Authentication/PersonaType'
 import types from '../../../../common/types'
 
 import { injectResponseFixtureIfFaked } from '../../../../common/cypress/integration/common/fakeServer'
