+++ cypress/e2e/Authentication/RegisterConsumer.feature
