+++ cypress/e2e/Authentication/RegisterConsumer/helpers.js
