+++ cypress/e2e/Authentication/RegisterConsumer/main.js
