--- cypress/integration/common/fakeServer.js
@@ -1,5 +1,0 @@
-export const injectResponseFixtureIfFaked = fixture => {
-  if (Cypress.env('fakeGraphql')) {
-    cy.fakeGraphqlResponseWithFixture(fixture)
-  }
-}
