--- cypress/integration/common/globalBefore.js
@@ -1,3 +1,0 @@
-before(() => {
-  cy.log('Will this run once before all tests?')
-})
