+++ cypress/integration/mytest.spec.js
@@ -1,0 +1,7 @@
+describe('my first test', function(){
+  context('my first context', function() {
+    it('can test', function () {
+      expect(true).to.be.false
+    })
+  })
+})
