--- cypress/support/stubHelpers.js
@@ -1,9 +1,0 @@
-export const responseStub = result => Promise.resolve({
-  json () {
-    return Promise.resolve(result)
-  },
-  text () {
-    return Promise.resolve(JSON.stringify(result))
-  },
-  ok: true
-})
