+++ package.json
@@ -10,11 +10,11 @@
     "build": "quasar build",
     "cypress:open:dev": "cross-env CYPRESS_baseUrl=http://localhost:4000/# cypress open",
     "cypress:open": "cross-env CYPRESS_baseUrl=http://localhost:4000/# cypress open",
-    "cypress:ci": "wait-on http://localhost:4000 && cross-env CYPRESS_baseUrl=http://localhost:4000/# cypress run --env configFile=acceptance",
-    "cypress:ci:dev": "yarn cypress:ci",
+    "cypress:integration:ci": "wait-on http://localhost:4000 && cross-env CYPRESS_baseUrl=http://localhost:4000/# cypress run --env configFile=integration",
+    "cypress:integration:ci:dev": "yarn cypress:integration:ci",
     "cypress:e2e": "wait-on http://localhost:4000 && cross-env CYPRESS_baseUrl=http://localhost:4000/# cypress run --env configFile=e2e",
-    "start:ci": "concurrently \"yarn cypress:ci\" --kill-others \"yarn start\" --success first",
-    "start:ci:dev": "concurrently \"yarn cypress:ci:dev\" --kill-others \"yarn start:dev\" --success first",
+    "start:integration:ci": "concurrently \"yarn cypress:integration:ci\" --kill-others \"yarn start\" --success first",
+    "start:integration:ci:dev": "concurrently \"yarn cypress:integration:ci:dev\" --kill-others \"yarn start:dev\" --success first",
     "start:e2e": "concurrently \"yarn cypress:e2e\" --kill-others \"yarn start\" --success first",
     "lint": "eslint --ext .js,.vue src",
     "start": "quasar serve dist/spa -H '0.0.0.0' -p 4000",
